-modified: find friends feature, now working as intended with no known bugs as of now

+++ ww_code/lib/pending_invites.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
@@ -12,6 +11,13 @@ class PendingInvitesPage extends StatefulWidget {
 }
 
 class PendingInvitesPageState extends State<PendingInvitesPage> {
+  int pendingInvitesCount = 0;
+
+  @override
+  void initState() {
+    super.initState();
+    fetchPendingInvitesCount();
+  }
 
   Future<void> _acceptFriendRequest(BuildContext context, String requestId) async {
     try {
@@ -28,7 +34,8 @@ class PendingInvitesPageState extends State<PendingInvitesPage> {
           .get();
 
       if (requestSnapshot.exists) {
-        Map<String, dynamic> requestData = requestSnapshot.data() as Map<String, dynamic>;
+        Map<String, dynamic> requestData =
+            requestSnapshot.data() as Map<String, dynamic>;
         String senderId = requestData['senderId'];
         String recipientId = requestData['recipientId'];
 
@@ -38,7 +45,22 @@ class PendingInvitesPageState extends State<PendingInvitesPage> {
 
         await _addFriendToCollection(senderId, recipientId);
         await _addFriendToCollection(recipientId, senderId);
-        
+
+        // Update friendsCount for both sender and recipient
+        await _incrementFriendsCount(senderId);
+        await _incrementFriendsCount(recipientId);
+
+        // Delete the friend request from recipient's PendingInvites
+        await FirebaseFirestore.instance
+            .collection('Users')
+            .doc(widget.userId)
+            .collection('PendingInvites')
+            .doc(requestId)
+            .delete();
+
+        // Update pending invites count
+        fetchPendingInvitesCount();
+
       }
     } catch (e) {
       ScaffoldMessenger.of(context).showSnackBar(
@@ -51,25 +73,87 @@ class PendingInvitesPageState extends State<PendingInvitesPage> {
     await FirebaseFirestore.instance.collection('Friends').add({
       'userId': userId,
       'friendId': friendId,
-      'status': 'accepted', 
+      'status': 'accepted',
     });
   }
 
   Future<void> _rejectFriendRequest(BuildContext context, String requestId) async {
     try {
+      // Get the friend request details to determine the recipient
+      DocumentSnapshot requestSnapshot = await FirebaseFirestore.instance
+          .collection('FriendRequests')
+          .doc(requestId)
+          .get();
+
+      if (!requestSnapshot.exists) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          const SnackBar(content: Text('Friend request not found!')),
+        );
+        return;
+      }
+
       // Delete the friend request
       await FirebaseFirestore.instance
           .collection('FriendRequests')
           .doc(requestId)
           .delete();
 
+      // Delete the friend request from recipient's PendingInvites
+      await FirebaseFirestore.instance
+          .collection('Users')
+          .doc(widget.userId)
+          .collection('PendingInvites')
+          .doc(requestId)
+          .delete();
+
+      // Update pending invites count
+      fetchPendingInvitesCount();
+
       ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Friend request rejected!')),
+        const SnackBar(content: Text('Friend request rejected')),
       );
     } catch (e) {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text('Error rejecting friend request: $e')),
       );
+    }
+  }
+
+  Future<void> fetchPendingInvitesCount() async {
+    try {
+      QuerySnapshot querySnapshot = await FirebaseFirestore.instance
+          .collection('Users')
+          .doc(widget.userId)
+          .collection('PendingInvites')
+          .get();
+
+      setState(() {
+        pendingInvitesCount = querySnapshot.size; // Update the count based on the query result
+      });
+    } catch (e) {
+      print('Error fetching pending invites count: $e');
+      // Handle error as needed
+    }
+  }
+
+  Future<void> _incrementFriendsCount(String userId) async {
+    try {
+      DocumentSnapshot userSnapshot = await FirebaseFirestore.instance
+          .collection('Users')
+          .doc(userId)
+          .get();
+
+      if (userSnapshot.exists) {
+        int currentFriendsCount = userSnapshot.get('friendsCount') ?? 0;
+        await userSnapshot.reference.update({
+          'friendsCount': currentFriendsCount + 1,
+        });
+
+    
+        
+      }
+    } catch (e) {
+      print('Error incrementing friends count for user $userId: $e');
     }
   }
 
@@ -93,16 +177,33 @@ class PendingInvitesPageState extends State<PendingInvitesPage> {
                 }
 
                 if (!snapshot.hasData || snapshot.data!.docs.isEmpty) {
-                  return const Center(child: Text('No pending invites'));
+                  return  Center(
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Image.asset(
+                          'images/empty_pending_invites.png',
+                          height: 200,
+                        ),
+                        const SizedBox(height: 16), 
+                        const Text('No pending invites',
+                        style: TextStyle(fontSize: 18)
+                        )
+                      ],
+                    ),                  
+                  );
                 }
 
                 return ListView(
                   children: snapshot.data!.docs.map((DocumentSnapshot doc) {
-                    Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
-                    String senderDisplayName = data['senderDisplayName'] ?? 'Unknown User';
+                    Map<String, dynamic> data =
+                        doc.data() as Map<String, dynamic>;
+                    String senderDisplayName =
+                        data['senderDisplayName'] ?? 'Unknown User';
                     String requestId = doc.id;
                     return ListTile(
-                      title: Text('$senderDisplayName sent you a friend request'),
+                      title:
+                          Text('$senderDisplayName sent you a friend request'),
                       trailing: Row(
                         mainAxisSize: MainAxisSize.min,
                         children: [
@@ -128,3 +229,4 @@ class PendingInvitesPageState extends State<PendingInvitesPage> {
     );
   }
 }
+
